requirements.txt
setuptools~=60.2.0
requests~=2.27.1
pytest~=7.1.1
PyYAML~=6.0